+++ disteval/basics/__init__.py
@@ -1,6 +1,4 @@
 # -*- coding:utf-8 -*-
-from __future__ import absolute_import, print_function, division
-
 from .classifier_characteristics import ClassifierCharacteristics
 from .preparation import prepare_data, shrink_data
 from .preparation import convert_and_remove_non_finites
