+++ disteval/visualization/__init__.py
@@ -4,7 +4,7 @@
 Collection of methods to visualize the results of disteval functions
 """
 from .feature_importance_test import visualize_feature_importance_mad
-from .roc_curve_equivalence_test import roc_curve_equivalence_ks_test
+from .roc_curve_equivalence_test import visualize_roc_curve_equivalence_test
 from .comparison_plotter import ComparisonPlotter
 
 __all__ = ['visualize_feature_importance_mad',
